% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addon-dplyr.R
\name{addon-dplyr}
\alias{addon-dplyr}
\alias{count_pct}
\alias{count_pct_group}
\title{Functions to summarise count with percentage within a dataframe}
\usage{
count_pct(.data, ..., sort = FALSE)

count_pct_group(.data, ...)
}
\value{
returns a dataframe with count and percentage columns
}
\description{
\code{count_pct} is the extenstion of dplyr::count with percent
\code{count_pct_group} is the extenstion of dplyr::count with percent by group
}
\examples{

library(dacol)
library(dplyr)
library(scales)

df = data.frame(x = c(1, 2, 3, 4, 5),
                y = c("b", "a", "k", "b", "k"),
                z = c(5, 8, 8, 6, 5))

df \%>\% count_pct(y)
df \%>\% count_pct(z)

mtcars \%>\% count_pct(cyl)
mtcars \%>\% count_pct(cyl, sort = TRUE)
mtcars \%>\% count_pct(cyl, am, sort = TRUE)

mtcars \%>\% count_pct_group(cyl)
mtcars \%>\% count_pct_group(cyl, am)
mtcars \%>\% count_pct_group(cyl, am) \%>\% arrange(desc(cyl))
}
