% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacol.R
\docType{package}
\name{dacol-package}
\alias{dacol}
\alias{dacol-package}
\title{dacol: Danai's utilities for adding columns}
\description{
dacol provides utilities to add/modify columns into dataframe.
}
\details{
The utilities include:

\itemize{
\item Statistical measures: mode, confident_interval
\item Data Transformation of a vector column: cosine, logistic, zscore
\item Distance between 2 vector columns: euclidean, pearson, cosine, canberra
\item Data Cleasing: trim_outlier
\item Percentile: decile, rank_percentile
}

To learn more about dplyr, start with the vignettes:
`browseVignettes(package = "dacol")`
}
\section{Package options}{

\describe{
\item{`dacol.show_progress`}{Should lengthy operations such as `do()`
  show a progress bar? Default: `TRUE`}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ldanai/dacol}
  \item Report bugs at \url{https://github.com/ldanai/dacol/issues}
}

}
\author{
\strong{Maintainer}: Danai L \email{ldanai@gmail.com}

}
